requirements.txt
streamlit
tensorflow-cpu
numpy
matplotlib
gwpy
pymongo
dnspython
boto3